update module fiels names
